+++ NEWS.md
@@ -1,3 +1,8 @@
+# worldbank 0.3.0
+
+- Moved to HTTPS for all API calls
+- Support for Poverty and Inequality Platform (PIP) data
+
 # worldbank 0.2.0
 
 - Support month and quarter date format in `wb_country_indicator()`